Moved backend.ts and frontend.ts up one level
