+++ src/tasks/start.ts
@@ -3,8 +3,8 @@ import { BuidlerPluginError } from '@nomiclabs/buidler/plugins'
 import { BuidlerRuntimeEnvironment } from '@nomiclabs/buidler/types'
 import { TASK_START } from './task-names'
 import { logMain } from '../ui/logger'
-import { startBackend } from './start/backend/backend'
-import { startFrontend } from './start/frontend/frontend'
+import { startBackend } from './start/backend'
+import { startFrontend } from './start/frontend'
 import { AragonConfig } from '~/src/types'
 import tcpPortUsed from 'tcp-port-used'
 import fsExtra from 'fs-extra'
